pyproject.toml
[project]
name = "ocr-clip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
    "pyperclip>=1.11.0",
    "pytesseract>=0.3.13",
]
